Move SHOW SCHEMAS LIKE tests to engine
Connector query tests could fail when concurrent schema changes made
SHOW SCHEMAS and SHOW SCHEMAS LIKE observe different listings. The
LIKE and ESCAPE behavior is engine-side, so exercise it in the engine
suite instead.
