+++ testing/trino-testing/src/main/java/io/trino/testing/AbstractTestQueries.java
@@ -22,14 +22,12 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.Set;
 
-import static com.google.common.base.Verify.verify;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static io.trino.connector.informationschema.InformationSchemaTable.INFORMATION_SCHEMA;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.VarcharType.VARCHAR;
 import static io.trino.testing.MaterializedResult.resultBuilder;
 import static io.trino.testing.QueryAssertions.assertContains;
-import static io.trino.testing.assertions.Assert.assertEventually;
 import static io.trino.tpch.TpchTable.NATION;
 import static io.trino.tpch.TpchTable.ORDERS;
 import static io.trino.tpch.TpchTable.REGION;
@@ -267,26 +265,6 @@ public abstract class AbstractTestQueries
     {
         MaterializedResult result = computeActual(format("SHOW SCHEMAS LIKE '%s'", getSession().getSchema().get()));
         assertThat(result.getOnlyColumnAsSet()).isEqualTo(ImmutableSet.of(getSession().getSchema().get()));
-    }
-
-    @Test
-    public void testShowSchemasLikeWithEscape()
-    {
-        assertQueryFails("SHOW SCHEMAS LIKE 't$_%' ESCAPE ''", "Escape string must be a single character");
-        assertQueryFails("SHOW SCHEMAS LIKE 't$_%' ESCAPE '$$'", "Escape string must be a single character");
-
-        // Using eventual assertion because set of schemas may change concurrently.
-        assertEventually(() -> {
-            Set<Object> allSchemas = computeActual("SHOW SCHEMAS").getOnlyColumnAsSet();
-            assertThat(allSchemas).isEqualTo(computeActual("SHOW SCHEMAS LIKE '%_%'").getOnlyColumnAsSet());
-            Set<Object> result = computeActual("SHOW SCHEMAS LIKE '%$_%' ESCAPE '$'").getOnlyColumnAsSet();
-            verify(allSchemas.stream().anyMatch(schema -> !((String) schema).contains("_")),
-                    "This test expects at least one schema without underscore in it's name. Satisfy this assumption or override the test.");
-            assertThat(result)
-                    .isSubsetOf(allSchemas)
-                    .isNotEqualTo(allSchemas);
-            assertThat(result).contains("information_schema").allMatch(schemaName -> ((String) schemaName).contains("_"));
-        });
     }
 
     @Test
